+++ Yolo/src/main/java/com/cfm/Yolo/repository/PersonRepository.java
@@ -3,8 +3,8 @@ package com.cfm.Yolo.repository;
 import com.cfm.Yolo.model.Person;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface PersonRepository extends JpaRepository<Person, Integer> {
-    
-    Person findPersonById(Integer id);
-    
+public interface PersonRepository extends JpaRepository<Person, Long> {
+
+  Person findPersonById(Long id);
+
 }
