+++ Yolo/src/main/java/com/cfm/Yolo/repository/UserRepository.java
@@ -4,9 +4,9 @@ import com.cfm.Yolo.model.User;
 
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserRepository extends JpaRepository<User, Integer> {
+public interface UserRepository extends JpaRepository<User, Long> {
 
-  User findUserById(Integer id);
+  User findUserById(Long id);
 
   User findUserByUsername(String username);
